requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# OCR
easyocr==1.7.1
pdf2image==1.17.0
Pillow==10.2.0

# LLM
google-generativeai==0.3.2

# Optional but useful
python-magic-bin==0.4.14  # For file type detection on Windows
